Move ProgramEnrolment directory one level up
